+++ pulsar_data_collection2/config.py
@@ -1,0 +1,3 @@
+import database_connectors as db
+
+factories = {"influxdb": db.Influxdb}
