+++ pulsar_data_collection2/database_connectors/database_actions.py
@@ -13,11 +13,11 @@ class DatabaseActions(ABC):
 
     @abstractmethod
     def make_connection(self):
-        pass
+        ...
 
     @abstractmethod
     def write_data(self):
-        pass
+        ...
 
 
 class DatabaseActionsFactory(ABC):
@@ -28,3 +28,4 @@ class DatabaseActionsFactory(ABC):
     @abstractmethod
     def get_database_actions(self) -> DatabaseActions:
         """Returns a new video exporter belonging to this factory."""
+        ...
