+++ pulsar_data_collection2/database_connectors/influxdb.py
@@ -1,8 +1,13 @@
 from database_actions import DatabaseActions, DatabaseActionsFactory
 
+# from influxdb import InfluxDBClient
+
 
 class InfluxdbActions(DatabaseActions):
     """Infuxdb database connection"""
+
+    def __init__(self):
+        pass
 
     def make_connection(self):
         """Makes connection to the Influxdb database"""
@@ -10,6 +15,7 @@ class InfluxdbActions(DatabaseActions):
 
     def write_data(self):
         """Sends data to the database"""
+        pass
 
 
 class Influxdb(DatabaseActionsFactory):
